+++ src/app/api/project/crewTaskHistory/route.ts
@@ -2,6 +2,7 @@ import { NextRequest } from 'next/server';
 import authApi from '@/app/api/_interceptor/authApi';
 import { routeResponse } from '@/app/api/_interceptor/routeResponse';
 
+// todo - api, route 주석제거
 /**
  * 프로젝트 크루 업무 이력 조회
  * @param req
@@ -15,6 +16,7 @@ export async function GET(req: NextRequest) {
 
   const res = await authApi(
     `api/projectmember/${projectMemberId}/works?pageIndex=${pageIndex}&itemCount=${itemCount}`,
+    { method: req.method },
   );
 
   return routeResponse(req, res);
