+++ src/constants/defaultSelectOptions.ts
@@ -1,3 +1,3 @@
 export const DEFAULT_POSITION_OPTION = { name: '포지션', value: 0n } as const;
 
-export const DEFAULT_CREW_OPTION = { name: '멤버 선택', value: '0' } as const;
+export const DEFAULT_CREW_OPTION = { name: '크루 선택', value: '0' } as const;
