+++ src/features/projectCrews/private/service/getProjectCrewList.ts
@@ -2,20 +2,19 @@ import { request } from '@/utils/clientApi/request';
 import { useSuspenseQuery } from '@tanstack/react-query';
 import { ResponseBody } from '@/types/responseBody';
 import { ProjectCrew } from '@/features/projectCrews/private/types';
+import { bigIntToString } from '@/shared/utils/stringUtils';
 
-export const getProjectCrewList = async ({
-  projectId,
-}: {
-  projectId: string | bigint;
-}): Promise<ResponseBody<Record<'projectMembers', ProjectCrew[]>>> => {
+export const getProjectCrewList = async (
+  projectId: bigint,
+): Promise<ResponseBody<Record<'projectMembers', ProjectCrew[]>>> => {
   return await request('GET', `/api/project/crews/list?projectId=${projectId}`);
 };
 
 export const CREW_LIST_QUERY_KEY = 'crewList';
 
-export const useProjectCrewList = (projectId: string) => {
+export const useProjectCrewList = (projectId: bigint) => {
   return useSuspenseQuery({
-    queryKey: [CREW_LIST_QUERY_KEY, projectId],
-    queryFn: () => getProjectCrewList({ projectId }),
+    queryKey: [CREW_LIST_QUERY_KEY, bigIntToString(projectId)],
+    queryFn: () => getProjectCrewList(projectId),
   });
 };
